backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic>=2.0
pydantic-settings
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart

# Fuerza una versión estable de bcrypt
bcrypt==4.1.2

# google required libraries
google-api-python-client
google-auth
